# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DevJobs",
        "description": "Documentación de la API para la plataforma DevJobs",
        "contact": {
            "name": "David",
            "email": "david@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://devjobs.test",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/genders": {
            "get": {
                "tags": [
                    "Genders"
                ],
                "summary": "Obtener todos los géneros",
                "description": "Requiere autenticación mediante token. Accesible solo para roles 'moderator' o 'god'.",
                "operationId": "7d27450f855b47fe95c650bf9ae8bee1",
                "responses": {
                    "200": {
                        "description": "Lista de géneros obtenida correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso denegado. Se requiere el rol 'moderator' o 'god'."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Genders"
                ],
                "summary": "Crear un nuevo género",
                "operationId": "a91a4df590ea0b9a75f375a3443a031c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Binario"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Género creado exitosamente"
                    },
                    "403": {
                        "description": "No autorizado: solo para roles 'moderator' o 'god'"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/genders/{id}": {
            "get": {
                "tags": [
                    "Genders"
                ],
                "summary": "Obtener un género por ID",
                "operationId": "b8d983ec4a7d75bd6cedbeded6761eac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Género encontrado"
                    },
                    "404": {
                        "description": "Género no encontrado"
                    },
                    "403": {
                        "description": "No autorizado: solo para roles 'moderator' o 'god'"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Genders"
                ],
                "summary": "Actualizar un género",
                "operationId": "fb8ceb72fd7d0e21faf597319f74b531",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Trinario"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Género actualizado correctamente"
                    },
                    "403": {
                        "description": "No autorizado: solo para roles 'moderator' o 'god'"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Genders"
                ],
                "summary": "Eliminar un género",
                "operationId": "22b5f1b4cc6cb46c6c1dd939f3f2672f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Género eliminado correctamente"
                    },
                    "403": {
                        "description": "No autorizado: solo para roles 'moderator' o 'god'"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/genders/{id}/toggle-default": {
            "post": {
                "tags": [
                    "Genders"
                ],
                "summary": "Alternar el estado por defecto del género",
                "operationId": "d5c5e19a9a6ea5f3e4cda48f0178f692",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del género",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado predeterminado actualizado correctamente"
                    },
                    "403": {
                        "description": "No autorizado: solo para rol 'god'"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}